requirements.txt
amqp==2.6.1
asgiref==3.2.10
async-generator==1.10
attrs==20.2.0
Babel==2.8.0
billiard==3.6.3.0
bleach==3.2.1
celery==4.4.7
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
defusedxml==0.6.0
Django==3.1.2
djangorestframework==3.12.1
entrypoints==0.3
flower==0.9.2
humanize==3.1.0
ipython-genutils==0.2.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
jupyterlab-pygments==0.1.2
kombu==4.6.11
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.1
nbconvert==5.3.1
nbformat==5.0.8
nest-asyncio==1.4.1
packaging==20.4
pandocfilters==1.4.2
prometheus-client==0.8.0
prompt-toolkit==3.0.8
psycopg2==2.8.6
psycopg2-binary==2.8.6
Pygments==2.7.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.1
pyzmq==19.0.2
six==1.15.0
sqlparse==0.4.1
testpath==0.4.4
tornado==4.2
traitlets==5.0.5
vine==1.3.0
wcwidth==0.2.5
webencodings==0.5.1
